Refactored all metrics stuff into its own module (#249)
