+++ pulsar-functions/conf/function_worker.yml
@@ -30,7 +30,7 @@ pulsarWebServiceUrl: http://localhost:8080
 numFunctionPackageReplicas: 1
 downloadDirectory: /tmp/pulsar_functions
 metricsConfig:
-  metricsSinkClassName: org.apache.pulsar.functions.runtime.metrics.sink.PrometheusSink
+  metricsSinkClassName: org.apache.pulsar.functions.metrics.sink.PrometheusSink
   metricsCollectionInterval: 30
   metricsSinkConfig:
     path: /metrics
