+++ pulsar-functions/metrics/src/main/java/org/apache/pulsar/functions/metrics/MetricsSink.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.functions.runtime.metrics;
+package org.apache.pulsar.functions.metrics;
 
 import java.util.Map;
