+++ pulsar-functions/metrics/src/test/java/org/apache/pulsar/functions/metrics/sink/PrometheusSinkTests.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.functions.runtime.metrics.sink;
+package org.apache.pulsar.functions.metrics.sink;
 
 import java.io.IOException;
 import java.util.Arrays;
